broke the Firestore component into two components to handle the updateDoc function correctly
